Got the PDF meta-data rendering in the template...
Fucking finally!!!
So, got rid of this unicode error that says something like
"ordinal not in range(128)"

Apparently Django has some nice stuff to deal with this.
The smart_str and smart_unicode functions from
 django.utils.encoding are used here for now to get something
going. Next stop: find a solution that doesn't require django
as a dependency for rendering one single html page. :P

The exact error:
 UnicodeEncodeError: 'ascii' codec can't encode character
 {some unicode char} in position {some position}: ordinal not
 in range(128)

With thanks to
<http://www.saltycrane.com/blog/2008/11/python-unicodeencodeerror-ascii-codec-cant-encode-character/>